Implemented room controller

+++ backend/boutique_app/src/main/kotlin/at/fhtw/model/repositories/RoomRepository.kt
@@ -3,4 +3,4 @@ package at.fhtw.model.repositories
 import at.fhtw.model.entities.Room
 import org.springframework.data.repository.CrudRepository
 
-interface RoomRepository: CrudRepository<Room, Int>
+interface RoomRepository: CrudRepository<Room, Long>
